Update BASH3_ConverterVEGASToGenesPvalues.R
Hi Folks,
I really don't know how to use github, so I ended up just overwriting this file with a version of the script that actually works.
The major changes were in the shebang line - apparently the env for the R library containing the updated version of dmGWAS and biocParallel needs to be specified later in the actual R script, after other code packages have been read in from the default directory. I also fixed the argument line (R requires that datatypes be specified), and added more explanation in response to the user requesting help. 
Megan