Update BASH4_dmGWAS.R
Hi Folks,
I really don't know how to use github, so I ended up just overwriting my script with a newer version that actually works.
The primary changes were in the shebang line - the path to the R library that holds the new version of dmGWAS and BioCParallel needs to be specified later after loading in other code packages from the default R library, as well as the argument lines (R requires that datatypes be specified), and one of the column names in the input changed from "Gene" to "gene" and this broke the for loop, so that is now updated.  I have also added better explanation in response to the user requesting help.