Improve confusing message at end of meta-net-var
The script always terminated with
    Could not run: [list of things that could not run because of prerequisites]

But that was easily mistaken for an error message when the script
finished. I made the message more verbose and clear.

Now, if everything had all its prereqs, it just says,

    Finished running analyzers

And if something was missing prereqs, it says,

    Some analyzers could not run because they were missing prerequisites: analyzer 1 is missing foo, bar,analyzer 2 is missing bar, baz
    Finished running analyzers

Because of technical limitatations, I am committing this before testing.