Change heatmap generation input from GENE_PVALUE to TABBED_GENE_LIST
It turned out that the file hotnet2 needed as input wasn't the
GENE_PVALUE association but just a tabbed list of genes. This is what
we were passing in from felix's directory command-line. And since it
was the only thing getting that file, no one noticed that it was the
wrong format until I wanted to compare it to the GENE_PVALUE file
produced from VEGAS (since the descriptions were similar but hotnet
did not mention the header that the vegas file had). I discovered they
were radically different and made a new filetype for the one used by
hotnet.

I changed the name of the gene_p_value command-line argument too so
that other users would find the error.

I have tested the meta-script after the change - everything seems to
work the same.